Added more time based queries and split them into individual files
